Resolve various MSVC10 compiler warnings in 32-bit and 64-bit mode.

+++ kdiff3/src-QT4/diff.cpp
@@ -2086,7 +2086,7 @@ bool fineDiff(
    int k1=0;
    int k2=0;
    bool bTextsTotalEqual = true;
-   int listSize = diff3LineList.size();
+   int listSize = (int) diff3LineList.size();
    int listIdx = 0;
    for( i= diff3LineList.begin(); i!= diff3LineList.end(); ++i)
    {
